+++ src/itensor.jl
@@ -674,6 +674,8 @@ function dense(A::ITensor)
   return setinds(itensor(dense(tensor(A))), removeqns(inds(A)))
 end
 
+removeqns(T::ITensor) = dense(T)
+
 denseblocks(D::ITensor) = itensor(denseblocks(tensor(D)))
 
 """
@@ -1071,6 +1073,14 @@ For example, for dense tensors this may return `1:length(A)`, while for sparse t
 it may return a Cartesian range.
 """
 eachindex(A::ITensor) = eachindex(tensor(A))
+
+"""
+    eachindval(A::ITensor)
+
+Create an iterable object for visiting each element of the ITensor `A` (including structually
+zero elements for sparse tensors) in terms of pairs of indices and values.
+"""
+eachindval(T::ITensor) = eachindval(inds(T))
 
 """
     iterate(A::ITensor, args...)
@@ -1541,6 +1551,67 @@ function combiner(; kwargs...)
   return itensor(Combiner(), ())
 end
 
+@doc """
+    combiner(inds::Indices; kwargs...)
+
+Make a combiner ITensor which combines the indices (of type Index)
+into a single, new Index whose size is the product of the indices 
+given. For example, given indices `i1,i2,i3` the combiner will have 
+these three indices plus an additional one whose dimension is the 
+product of the dimensions of `i1,i2,i3`.
+
+Internally, a combiner ITensor uses a special storage type which
+means it does not hold actual tensor elements but just information
+about how to combine the indices into a single Index. Taking a product
+of a regular ITensor with a combiner uses special fast algorithms to
+combine the indices.
+
+To obtain the new, combined Index that the combiner makes out of
+the indices it is given, use the `combinedind` function.
+
+To undo or reverse the combining process, uncombining the Index back
+into the original ones, contract the tensor having the combined Index
+with the conjugate or `dag` of the combiner. (If the combiner is an ITensor 
+`C`, multiply by `dag(C)`.)
+
+### Example
+```
+# Combine indices i and k into a new Index ci
+T = randomITensor(i,j,k)
+C = combiner(i,k)
+CT = C * T
+ci = combinedind(C)
+
+# Uncombine ci back into i and k
+TT = dag(C) * CT
+
+# TT will be the same as T
+@show norm(TT - T) ≈ 0.0
+```
+         
+              i  j  k
+              |  |  |
+     T   =    =======
+
+              ci  i  k
+              |   |  |
+     C   =    ========
+              
+              ci  j
+              |   |
+     C * T =  =====
+
+""" combiner
+
+"""
+    combinedind(C::ITensor)
+
+Given a combiner ITensor, return the Index which is
+the "combined" index that is made out of merging
+the other indices given to the combiner when it is made
+
+For more information, see the `combiner` function.
+"""
 function combinedind(T::ITensor)
   if storage(T) isa Combiner && order(T) > 0
     return inds(T)[1]