+++ src/qn/qnindex.jl
@@ -8,6 +8,8 @@ qn(qnblock::QNBlock) = qnblock.first
 # Get the dimension of the specified block
 blockdim(qnblock::QNBlock) = qnblock.second
 
+NDTensors.resize(qnblock::QNBlock, newdim::Int64) = QNBlock(qnblock.first, newdim)
+
 # Get the dimension of the specified block
 blockdim(qnblocks::QNBlocks, b::Integer) = blockdim(qnblocks[b])
 blockdim(qnblocks::QNBlocks, b::Block{1}) = blockdim(qnblocks[only(b)])
@@ -35,12 +37,40 @@ function (qn1::QNBlock + qn2::QNBlock)
   return QNBlock(qn(qn1), blockdim(qn1) + blockdim(qn2))
 end
 
+function removeqn(qn_block::QNBlock, qn_name::String)
+  return removeqn(qn(qn_block), qn_name) => blockdim(qn_block)
+end
+
 function -(qns::QNBlocks)
   qns_new = copy(qns)
   for i in 1:length(qns_new)
     qns_new[i] = -qns_new[i]
   end
   return qns_new
+end
+
+function mergeblocks(qns::QNBlocks)
+  qnsC = [qns[1]]
+
+  # Which block this is, after combining
+  block_count = 1
+  for i in 2:nblocks(qns)
+    if qn(qns[i]) == qn(qns[i - 1])
+      qnsC[block_count] += qns[i]
+    else
+      push!(qnsC, qns[i])
+      block_count += 1
+    end
+  end
+  return qnsC
+end
+
+function removeqn(space::QNBlocks, qn_name::String; mergeblocks=true)
+  space = QNBlocks([removeqn(qn_block, qn_name) for qn_block in space])
+  if mergeblocks
+    space = ITensors.mergeblocks(space)
+  end
+  return space
 end
 
 """
@@ -138,6 +168,21 @@ end
 
 dim(i::QNIndex) = dim(space(i))
 
+"""
+    nblocks(i::QNIndex)
+
+Returns the number of QN blocks, or subspaces, of the QNIndex `i`.
+
+To obtain the dimension of block number `b`, use `blockdim(i,b)`.
+To obtain the QN associated with block `b`, use `qn(i,b)`.
+
+### Example
+```
+julia> i = Index([QN("Sz",-1)=>2, QN("Sz",0)=>4, QN("Sz",1)=>2], "i")
+julia> nblocks(i)
+3
+```
+"""
 nblocks(i::QNIndex) = nblocks(space(i))
 # Define to be 1 for non-QN Index
 nblocks(i::Index) = 1
@@ -171,6 +216,22 @@ qn(i::QNIndex, b::Block{1}) = qn(space(i), b)
 qn(ib::Pair{<:Index,Block{1}}) = qn(first(ib), last(ib))
 
 # XXX: deprecate the Integer version
+# Miles asks: isn't it pretty convenient to have it?
+"""
+    qn(i::QNIndex, b::Integer)
+
+Returns the QN associated with block number `b` of 
+a QNIndex `i`.
+
+### Example
+```
+julia> i = Index([QN("Sz",-1)=>2, QN("Sz",0)=>4, QN("Sz",1)=>2], "i")
+julia> qn(i,1)
+QN("Sz",-1)
+julia> qn(i,2)
+QN("Sz",0)
+```
+"""
 qn(i::QNIndex, b::Integer) = qn(i, Block(b))
 
 # Get the QN of the block the IndexVal lies in
@@ -191,7 +252,24 @@ end
 qnblocks(i::QNIndex) = space(i)
 
 # XXX: deprecate the Integer version
+# Miles asks: isn't the integer version very convenient?
 blockdim(i::QNIndex, b::Block) = blockdim(space(i), b)
+
+"""
+    blockdim(i::QNIndex, b::Integer)
+
+Returns the dimension of block number `b` of
+a QNIndex `i`.
+
+### Example
+```
+julia> i = Index([QN("Sz",-1)=>2, QN("Sz",0)=>4, QN("Sz",1)=>2], "i")
+julia> blockdim(i,1)
+2
+julia> blockdim(i,2)
+4
+```
+"""
 blockdim(i::QNIndex, b::Integer) = blockdim(i, Block(b))
 function blockdim(i::Index, b::Union{Block,Integer})
   return error(
@@ -357,6 +435,8 @@ end
 # Make a new Index with the specified qn blocks
 replaceqns(i::QNIndex, qns::QNBlocks) = setspace(i, qns)
 
+NDTensors.block(i::QNIndex, n::Integer) = space(i)[n]
+
 function setblockdim!(i::QNIndex, newdim::Integer, n::Integer)
   qns = space(i)
   qns[n] = qn(qns[n]) => newdim
@@ -366,6 +446,12 @@ end
 function setblockqn!(i::QNIndex, newqn::QN, n::Integer)
   qns = space(i)
   qns[n] = newqn => blockdim(qns[n])
+  return i
+end
+
+function setblock!(i::QNIndex, b::QNBlock, n::Integer)
+  qns = space(i)
+  qns[n] = b
   return i
 end
 
@@ -386,6 +472,10 @@ function combineblocks(i::QNIndex)
 end
 
 removeqns(i::QNIndex) = setdir(setspace(i, dim(i)), Neither)
+function removeqn(i::QNIndex, qn_name::String; mergeblocks=true)
+  return setspace(i, removeqn(space(i), qn_name; mergeblocks))
+end
+mergeblocks(i::QNIndex) = setspace(i, mergeblocks(space(i)))
 
 function addqns(i::Index, qns::QNBlocks; dir::Arrow=Out)
   @assert dim(i) == dim(qns)
@@ -427,6 +517,8 @@ hassameflux(::Index, ::QNIndex) = false
 
 # Split the blocks into blocks of size 1 with the same QNs
 splitblocks(i::Index) = setspace(i, splitblocks(space(i)))
+
+trivial_space(i::QNIndex) = [QN() => 1]
 
 function mutable_storage(::Type{Order{N}}, ::Type{IndexT}) where {N,IndexT<:QNIndex}
   return SizedVector{N,IndexT}(undef)